+++ src/verify-code/index.ts
@@ -22,6 +22,7 @@ Component({
     enableNative: false, // 兼容安卓input的输入bug
     countDown: 60,
     isInitialActive: true,
+    initActive: false, // 是否自动触发点击发送事件
     onInput: () => {},
     onConfirm: () => {},
     onFocus: () => {},
@@ -45,6 +46,14 @@ Component({
       _countDown: this.props.countDown,
       actedBefore: false,
     });
+    // 在组件加载的时候是否主动触发点击发送验证码事件
+    if (this.props.initActive) {
+      this.noSendCountDown();
+    } else {
+      this.setData({
+        _actionActive: !this.props.initActive,
+      });
+    }
   },
   didUnmount() {
     clearInterval(this._timeout);
@@ -100,6 +109,30 @@ Component({
         }, 1000);
         const event = fmtEvent(this.props, e);
         onSend(event);
+      }
+    },
+    noSendCountDown() {
+      const { countDown } = this.props;
+      if (this.data._actionActive) {
+        this.setData({
+          _actionActive: false,
+        });
+        this._timeout = setInterval(() => {
+          const subOne = this.data._countDown - 1;
+          if (subOne <= 0) {
+            clearInterval(this._timeout);
+            this.setData({
+              _actionActive: true,
+              resend: true,
+              _countDown: countDown,
+              actedBefore: true,
+            });
+          } else {
+            this.setData({
+              _countDown: subOne,
+            });
+          }
+        }, 1000);
       }
     },
   },